Fix BEP7 compatibility with IPv6 trackers and IPv4 peers
Currently the BEP7 code in tracker_http supports the case where
a peer is IPv6 ready but the tracker is not, passing its IPv6 IP
as an announce parameter. Unfortunately it does not yet support
the case where both the peer and tracker are running in IPv6 mode
but other peers in the swarm are not yet IPv6 capable. This sort
of situation is occuring with increasing frequency now that IPv6
adoption is taking off.

To correct this, we are now going to send the IPv4 parameter to
the tracker when connecting to the tracker over IPv6, implementing
the remaining portion of this specification amendment.