Drop packets everywhere we're deleting DhtTransaction's (#135)
This is a continuation of 856d9ca810d437656c526d40cdded14d095e6695

I've observed at least two more behaviours where this deletion of
the DhtTransaction within dht_server.cc can later lead to a use
after free when attempting to process the associated packets.

As such this commit updates all of the areas where we're deleting
DhtTransactions to also drop the corresponding packets. While this
arguably could be placed within the DhtTransactions' destructor, it
feels like that would be messing up the hierarchy within the Dht*
classes.