[RFC] Use native alignment and the standard allocator
The cacheline allocator and alignment hints passed to the compiler
rely upon /usr/include/linux/cache.h, a file which is no longer
present on any linux distro I checked. This resulted in it using
the safe fallback of 128 byte alignment.

Unfortunately, 128 byte alignment is oversized for a wide range of
structures and types we're enforcing this alginment on. Furthermore
we're occasionally passing these properties to functions that are not
expecting the 128 byte alignment.

UBSan in Clang 3.7 found ~230 sections where 128 byte aligned objects
were handled by code that wasn't expecting them to be 128 byte aligned.
This doesn't appear to cause issues in practice because the generous
padding is simply ignored.

The same issue exists within rtorrent as well, and a separate pull
request for it will be created at a later date if this is approved.