Align torrent block vectors
This pull request implements memory alignment for a few vectors with the custom `rak::cacheline_allocator` class.

It improves download throughput by approximately 25-30%, with a significant reduction in CPU usage. It's a hot path to the torrent download delegator, which uses the majority of the CPU usage when downloading torrents.

We're no longer allowed to delete some Block constructors because we're implementing the `rak::cacheline_allocator` class.  This code has been commented out.