Fix undefined behavior in extents::insert
The extents::insert may dereference the end iterator, which is undefined
behavior. This occurs when there is no element that is greater than the
specified key. In which case, iter will point to the element before the
end pointer. The loop body will then increment the iter (causing it to
point to the end pointer) and dereference it, unless there is an
overlap.

The "borders" test case from `test_extents.cc` triggers this behavior
and causes test failures on some architectures. You can single-step
through it with GDB to confirm this, Specifically, this was encountered
on the Alpine Linux Edge armhf/x86 builders.

If I am understand the semantics correctly, then if we ran into the end
iterator at this point, there can be no overlap and therefore we must
insert the element.